Always return a `PyObject` if requested type is `PyObject`.
Adds a higher-priority check if the requested type is PyObject. If it is, a PyObject will be returned.

Fixes an issue with numpy arrays – currently, they cannot be retrieved as a `PyObject`, and attempting to do so results in a `TypeError` (since numpy conversion logic has no escape hatch if expected type is not a Java array nor Java `NDArray`)