Add a link to CONTRIBUTION.md to the root README.rst
Since `CONTRIBUTION.md` is not in the root of the project, novice
GitHub users may not know to look in the `.github` folder to find
it.

Adding a link to `CONTRIBUTION.md` to the root `README.rst` gives
more visibility to these guidelines.

For those that want to know more about `CONTRIBUTION.md`, see
https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors .