Fix usages of PyThreadState with the limited ABI
When the limited ABI is enabled, we can't access `PyThreadState.interp`.
Instead, we have to use the API function `PyThreadState_GetInterpreter`.

Unfortunately, this API is new as of Python 3.9 so versions
before that simply cannot know the current interpreter that
is running in the main thread......
As such, I include a new define `JEP_ASSUME_SINGLE_INTERPRETER`
where we just kind of wing it and assume there is only one active
interpreter (which is actually pretty common in
CPython land - a lot of extensions aren't built with subinterpreters
in mind).